% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/multiStepPage.R
\name{changeStep}
\alias{changeStep}
\title{Changes current step}
\usage{
changeStep(session, id, page)
}
\arguments{
\item{session}{}

\item{id}{}

\item{page}{}
}
\description{
Changes current step to the one given
}

