% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/multiStepPage.R
\name{multiStepPage}
\alias{multiStepPage}
\title{MultiStepPage element}
\usage{
multiStepPage(id, ..., title = NULL, topButtons = FALSE,
  bottomButons = TRUE)
}
\arguments{
\item{id}{}

\item{title}{}

\item{topButtons}{}

\item{\dots}{}

\item{bottomButtons}{}

\item{width}{}

\item{height}{}
}
\value{
A multistep page container (div)
}
\description{
Creates a multistep page element for shiny
}

