% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiStepPage.R
\name{multiStepPageUpdateHeight}
\alias{multiStepPageUpdateHeight}
\title{Force page heigth update (for environments liketabpages)}
\usage{
multiStepPageUpdateHeight(session, id)
}
\arguments{
\item{id}{}
}
\description{
Force page heigth update (for environments liketabpages)
}

