% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/multiStepPage.R
\name{stepPage}
\alias{stepPage}
\title{Create a page for a multStepPage element}
\usage{
stepPage(description, ...)
}
\arguments{
\item{description}{A short description of the page for the tracking box}

\item{...}{Page content}
}
\description{
Creates a single page for a multistep element. \code{\link{multiStepPage}}. By default shiny doesnt render hidden outputs
(kind of lazy web-evaluation). To ensure that all animations work propperly use \code{\link{outputOptions}} with
the \code{suspendWhenHidden} option over all outputs inside a step page
}

