% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiStepPage.R
\name{multiStepPage}
\alias{multiStepPage}
\title{MultiStepPage element}
\usage{
multiStepPage(id, ..., title = NULL, topButtons = FALSE,
  bottomButons = TRUE)
}
\arguments{
\item{height}{}
}
\value{
A multistep page container (div)
}
\description{
Creates a multistep page element for shiny
}

