% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/multiStepPage.R
\name{switchStepBack}
\alias{switchStepBack}
\title{Allows/disallows going back}
\usage{
switchStepBack(session, id, status)
}
\arguments{
\item{session}{}

\item{id}{}

\item{status}{(T/F)}
}
\description{
Allows/disallows going back
}

